Module oak_debug_service::proto::perftools::profiles

Structs

Function

Label

Line

Location

Describes function and line table debug information.

Mapping

Profile

Sample

Each Sample records values encountered in some program context. The program context is typically a stack trace, perhaps augmented with auxiliary information like the thread-id, some indicator of a higher level request being handled etc.

ValueType

ValueType describes the semantics and measurement units of a value.